+++ tests/files/util/mod.rs
@@ -69,7 +69,8 @@ macro_rules! set_artist {
 
 		$file_write.rewind().unwrap();
 
-		$tag.save_to(&mut $file_write).unwrap();
+		$tag.save_to(&mut $file_write, lofty::WriteOptions::new())
+			.unwrap();
 	};
 }